Fix EagleTime usize→u64: prevent overflow on 32-bit WASM targets
EtType::u and EtType::i were using usize/isize which truncates to 32 bits
on WASM32, overflowing Eagle Time oscillation counts (~2.5e18 since 1969).
Changed to u64/i64 for correct cross-platform behavior.

Built with Claude Opus 4.6

+++ src/decoding/metadata.rs
@@ -68,11 +68,11 @@ pub fn parse_eagle_time(data: &[u8], pointer: &mut usize) -> Result<VsfType, Err
     match time_type {
         b'u' => {
             let value = decode_usize(data, pointer)?;
-            Ok(VsfType::e(EtType::u(value)))
+            Ok(VsfType::e(EtType::u(value as u64)))
         }
         b'i' => {
             let value = decode_isize(data, pointer)?;
-            Ok(VsfType::e(EtType::i(value)))
+            Ok(VsfType::e(EtType::i(value as i64)))
         }
         b'f' => {
             // Eagle Time floats: [e][f][5 or 6][bytes]
